Move drag image generation to checkers/drag.py
